refactor: reorganize time tracking DTOs into dedicated package
