+++ backend/src/main/java/org/pertitrack/backend/dto/timeTrackingDto/TimeRecordResponse.java
@@ -1,4 +1,4 @@
-package org.pertitrack.backend.dto;
+package org.pertitrack.backend.dto.timeTrackingDto;
 
 import lombok.With;
 import org.pertitrack.backend.entity.timetrack.TimeRecord;
@@ -7,12 +7,19 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @With
-public record TimeRecordUpdateRequest(
+public record TimeRecordResponse(
+        String id,
+        String employeeId,
+        String employeeFirstName,
+        String employeeLastName,
         LocalDate recordDate,
         LocalDateTime recordTime,
         TimeRecord.RecordType recordType,
         TimeRecord.LocationType locationType,
         String notes,
-        Boolean isManual
+        Boolean isManual,
+//        Boolean isApproved,
+        LocalDateTime createdAt,
+        LocalDateTime updatedAt
 ) {
 }
